added impiegato, fattura, fornitore, ore

+++ src/main/java/it/contrader/model/Impiegato.java
@@ -58,7 +58,7 @@ public class Impiegato {
 	
 	@NotNull
 	@Column(name = "costo_lordo_annuo")
-	private int costoLordoAnnuo;
+	private double costoLordoAnnuo;
 	
 	@OneToMany(mappedBy = "impiegato")
 	@OnDelete(action = OnDeleteAction.CASCADE)
